arm64: apple: add bindings for gpio, spi gate clocks, and keyboard
This is 3c945561b6597c6b86ec93e82f8eddfa995b7299 from
corellium/linux-m1, but applied to a different file.

Now that we have a driver for gate clocks add the proper topology for the
UART. These are already enabled by the bootloader but are part of the
clock topology used by devices yet to be implemented.

Signed-off-by: Joey Gouly <joey.gouly@arm.com>
Signed-off-by: Sven Peter <sven@svenpeter.dev>

[joey: cherry picked db975d0b41911068768f0bb23998e2b7d4d59807 from
AsahiLinux/linux]