PCI: apple: Add driver for the Apple M1
Add a driver for the PCIe controller found in Apple system-on-chips,
particularly the Apple M1. This driver exposes the internal bus used for
the USB type-A ports, Ethernet, Wi-Fi, and Bluetooth. This patch brings
up the USB type-A ports and Ethernet. Bringing up the radios requires a
driver for the System Management Coprocessor, which in turn requires a
driver for Apple's bespoke mailbox protocol.

The driver consists of two major parts: hardware initialization and MSI
handling. The hardware initialization is derived from Mark Kettenis's
U-Boot patches which in turn is derived from Corellium's patches for the
hardware. The rest of the driver is derived from Marc Zyngier's driver
for the hardware.

Co-authored-by: Stan Skowronek <stan@corellium.com>
Signed-off-by: Stan Skowronek <stan@corellium.com>
Co-authored-by: Marc Zyngier <maz@kernel.org>
Signed-off-by: Marc Zyngier <maz@kernel.org>
Co-authored-by: Mark Kettenis <kettenis@openbsd.org>
Co-authored-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>
Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>