PCI: of: Allow matching of an interrupt-map local to a pci device
Just as we now allow an interrupt map to be parsed when part
of an interrupt controller, there is no reason to ignore an
interrupt map that would be part of a pci device node such as
a root port since we already allow interrupt specifiers.

This allows the device itself to use the interrupt map for
for its own purpose.

Signed-off-by: Marc Zyngier <maz@kernel.org>