brcmfmac: Forward MAC from device tree to NVRAM
On some systems, the preferred MAC address is specified in the
device-tree as a local-mac-addr node on the device. It is therefore
required to get this MAC address with of_get_mac_addr, and set it with
NVRAM. Add support for this forwarding.

This will be used for Wi-Fi on Apple chips, where the MAC address is
ultimately controlled by the first-stage bootloader and passed in the
Apple Device Tree. The Asahi Linux loader forwards this property to the
Linux Device Tree. It is up to us to forward it finally to the hardware.

Co-developed-by: Stan Skowronek <stan@corellium.com>
Signed-off-by: Stan Skowronek <stan@corellium.com>
Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>