arm64: dts: apple: j274: Expose PCI node for the Wi-Fi MAC address
This allows the bootloader to pass the MAC address through the device
tree, and in turn will allow the Wi-Fi driver to set the MAC address
appropriately. The usual mechanisms of MAC address fetching do not work
on Apple's variant of the Broadcom chip -- we need the information in
the device tree.

m1n1 already supports this if the property exists.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>