macsmd-hid: Respond to power button on Mini
On an M1 Mac Mini with 12.3 system firmware, the initial tap of the
power button is reported as button 0x01 rather than 0x06. Handle both
button IDs as power buttons. With this change, tapping the power button
on the Mini correctly triggers the power off dialog in GNOME and
triggers an immediate shutdown in fbcon, like the power button does on
other Linux machines as well as presumably the M1 MacBooks.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>